refactor evm into new module
