+++ src/eth1spec/evm/error.py
@@ -1,0 +1,5 @@
+class StackUnderflowError(Exception):
+    pass
+
+class StackOverflowError(Exception):
+    pass
