+++ src/eth1spec/evm/gas.py
@@ -1,0 +1,1 @@
+GAS_VERY_LOW = 3
